+++ src/include/duckdb/common/types/vector.hpp
@@ -22,6 +22,12 @@ class VectorListBuffer;
 struct SelCache;
 enum class VectorConstructorAction;
 
+template <class T>
+class VectorIterationHelper;
+template <class T>
+class VectorScanEntriesHelper;
+class VectorValidityHelper;
+
 //! Vector of values of a specified PhysicalType.
 class Vector {
 	friend struct ConstantVector;
@@ -38,14 +44,6 @@ class Vector {
 
 	friend class DataChunk;
 	friend class VectorCacheBuffer;
-
-	template <class T>
-	class VectorIterationHelper;
-
-	template <class T>
-	class VectorScanEntriesHelper;
-
-	class VectorValidityHelper;
 
 public:
 	//! Create a vector that slices another vector
@@ -199,18 +197,12 @@ public:
 	static void DebugShuffleNestedVector(Vector &vector, idx_t count);
 
 	template <class T>
-	VectorIterationHelper<T> ScanAllValues(idx_t count) const {
-		return VectorIterationHelper<T>(*this, count);
-	}
+	VectorIterationHelper<T> ScanAllValues(idx_t count) const;
 
 	template <class T>
-	VectorScanEntriesHelper<T> ScanValues(idx_t count) const {
-		return VectorScanEntriesHelper<T>(*this, count);
-	}
+	VectorScanEntriesHelper<T> ScanValues(idx_t count) const;
 
-	VectorValidityHelper ScanValidity(idx_t count) const {
-		return VectorValidityHelper(*this, count);
-	}
+	VectorValidityHelper ScanValidity(idx_t count) const;
 
 private:
 	//! Returns the [index] element of the Vector as a Value.
@@ -239,246 +231,6 @@ protected:
 	//! The buffer holding auxiliary data of the vector
 	//! e.g. a string vector uses this to store strings
 	mutable buffer_ptr<VectorBuffer> auxiliary;
-
-private:
-	class VectorValidityHelper {
-	public:
-		VectorValidityHelper(const Vector &vector, idx_t count) : count(count) {
-			vector.ToUnifiedFormat(count, format);
-		}
-
-		bool IsValid(idx_t i) const {
-			return format.validity.RowIsValid(format.sel->get_index(i));
-		}
-		bool CanHaveNull() const {
-			return !format.validity.AllValid();
-		}
-		idx_t size() const {
-			return count;
-		}
-
-	private:
-		UnifiedVectorFormat format;
-		idx_t count;
-	};
-
-	template <class T>
-	class VectorIterationHelper {
-	public:
-		VectorIterationHelper(const Vector &vector, idx_t count) : count(count) {
-			vector.ToUnifiedFormat(count, format);
-			data = UnifiedVectorFormat::GetData<T>(format);
-		}
-
-	private:
-		UnifiedVectorFormat format;
-		const T *data;
-		idx_t count;
-
-	private:
-		struct VectorValueEntry {
-			idx_t index;
-			T value;
-			bool is_valid;
-
-			bool IsValid() const {
-				return is_valid;
-			}
-		};
-
-		class VectorIterator {
-		public:
-			explicit VectorIterator(UnifiedVectorFormat &format, const T *data, idx_t index)
-			    : format(format), data(data), index(index) {
-			}
-
-		public:
-			VectorIterator &operator++() { // NOLINT: match stl API
-				++index;
-				return *this;
-			}
-			VectorIterator operator++(int) { // NOLINT: match stl API
-				auto tmp = *this;
-				++index;
-				return tmp;
-			}
-			VectorIterator &operator--() { // NOLINT: match stl API
-				--index;
-				return *this;
-			}
-			VectorIterator &operator+=(idx_t n) {
-				index += n;
-				return *this;
-			}
-			VectorIterator &operator-=(idx_t n) {
-				index -= n;
-				return *this;
-			}
-			VectorIterator operator+(idx_t n) const {
-				return VectorIterator(format, data, index + n);
-			}
-			VectorIterator operator-(idx_t n) const {
-				return VectorIterator(format, data, index - n);
-			}
-			int64_t operator-(const VectorIterator &other) const {
-				return static_cast<int64_t>(index) - static_cast<int64_t>(other.index);
-			}
-			bool operator==(const VectorIterator &other) const {
-				return index == other.index;
-			}
-			bool operator!=(const VectorIterator &other) const {
-				return index != other.index;
-			}
-			bool operator<(const VectorIterator &other) const {
-				return index < other.index;
-			}
-			bool operator<=(const VectorIterator &other) const {
-				return index <= other.index;
-			}
-			bool operator>(const VectorIterator &other) const {
-				return index > other.index;
-			}
-			bool operator>=(const VectorIterator &other) const {
-				return index >= other.index;
-			}
-			VectorValueEntry operator*() const {
-				return GetEntry(index);
-			}
-			VectorValueEntry operator[](idx_t n) const {
-				return GetEntry(index + n);
-			}
-
-		private:
-			VectorValueEntry GetEntry(idx_t i) const {
-				VectorValueEntry result;
-				result.index = i;
-				auto sel_idx = format.sel->get_index(i);
-				result.is_valid = format.validity.RowIsValid(sel_idx);
-				if (result.is_valid) {
-					result.value = data[sel_idx];
-				}
-				return result;
-			}
-
-		private:
-			UnifiedVectorFormat &format;
-			const T *data;
-			idx_t index;
-		};
-
-	public:
-		VectorIterator begin() { // NOLINT: match stl API
-			return VectorIterator(format, data, 0);
-		}
-		VectorIterator end() { // NOLINT: match stl API
-			return VectorIterator(format, data, count);
-		}
-		idx_t size() const {
-			return count;
-		}
-		VectorValueEntry operator[](idx_t i) const {
-			VectorValueEntry result;
-			result.index = i;
-			const auto sel_idx = format.sel->get_index(i);
-			result.is_valid = format.validity.RowIsValid(sel_idx);
-			if (result.is_valid) {
-				result.value = data[sel_idx];
-			}
-			return result;
-		}
-		//! Returns the value at the specified location without checking the NULL mask
-		T GetValueUnsafe(idx_t i) const {
-			return data[format.sel->get_index(i)];
-		}
-		bool CanHaveNull() const {
-			return !format.validity.AllValid();
-		}
-	};
-
-	template <class T>
-	class VectorScanEntriesHelper {
-	public:
-		VectorScanEntriesHelper(const Vector &vector, idx_t count) : count(count) {
-			vector.ToUnifiedFormat(count, format);
-			data = UnifiedVectorFormat::GetData<T>(format);
-		}
-
-	public:
-		struct VectorValueEntry {
-			idx_t index;
-			T value;
-		};
-
-	private:
-		class VectorScanIterator {
-		public:
-			explicit VectorScanIterator(UnifiedVectorFormat &format, const T *data, idx_t index, idx_t count)
-			    : format(format), data(data), count(count), can_have_null(!format.validity.AllValid()) {
-				r.index = index;
-				AdvanceToValid();
-			}
-
-		public:
-			VectorScanIterator &operator++() {
-				++r.index;
-				AdvanceToValid();
-				return *this;
-			}
-			VectorScanIterator operator++(int) {
-				auto tmp = *this;
-				++(*this);
-				return tmp;
-			}
-			bool operator!=(const VectorScanIterator &other) const {
-				return r.index != other.r.index;
-			}
-			const VectorValueEntry &operator*() const {
-				return r;
-			}
-
-		private:
-			void AdvanceToValid() {
-				if (!can_have_null) {
-					if (r.index < count) {
-						// we know this value is valid
-						r.value = data[format.sel->get_index(r.index)];
-					}
-					return;
-				}
-				for (; r.index < count; r.index++) {
-					auto idx = format.sel->get_index(r.index);
-					if (format.validity.RowIsValid(idx)) {
-						// found a valid value - stop
-						r.value = data[idx];
-						break;
-					}
-				}
-			}
-
-		private:
-			UnifiedVectorFormat &format;
-			const T *data;
-			VectorValueEntry r;
-			idx_t count;
-			bool can_have_null;
-		};
-
-	public:
-		VectorScanIterator begin() { // NOLINT: match stl API
-			return VectorScanIterator(format, data, 0, count);
-		}
-		VectorScanIterator end() { // NOLINT: match stl API
-			return VectorScanIterator(format, data, count, count);
-		}
-		idx_t size() const {
-			return count;
-		}
-
-	private:
-		UnifiedVectorFormat format;
-		const T *data;
-		idx_t count;
-	};
 };
 
 //! The VectorChildBuffer holds a child Vector
@@ -499,3 +251,5 @@ public:
 };
 
 } // namespace duckdb
+
+#include "duckdb/common/vector/vector_iterator.hpp"
